refactor : eventHandler 객체 추가가
