--- src/changeModalContents.js
@@ -1,35 +1,0 @@
-import RestaurantCard from "./components/RestaurantCard";
-import RestaurantDetailButtonContainer from "./components/RestaurantDetailButtonContainer";
-import storage from "./domain/storage.ts";
-import renderFilteredRestaurants from "./renderFilteredRestaurant";
-import { $ } from "./utils/dom";
-
-const changeModalContents = (restaurant, restaurantList) => {
-  const restaurantDetailModal = $(".restaurant-detail-modal");
-  restaurantDetailModal.innerHTML = "";
-
-  restaurantDetailModal.appendChild(
-    RestaurantCard(restaurant, {
-      clickFavorite: () => {
-        storage.saveRestaurantList(restaurantList.value);
-        renderFilteredRestaurants(restaurantList);
-      },
-    })
-  );
-
-  restaurantDetailModal.appendChild(
-    RestaurantDetailButtonContainer(restaurant, () => {
-      clickDelete(restaurant, restaurantList);
-    })
-  );
-};
-
-const clickDelete = (restaurant, restaurantList) => {
-  restaurantList.delete(restaurant);
-
-  storage.saveRestaurantList(restaurantList.value);
-
-  renderFilteredRestaurants(restaurantList);
-};
-
-export default changeModalContents;
